\name{shape_hydrolineClip}
\alias{shape_hydrolineClip}
\docType{data}
\title{Hydrology lines shapefile for Great Lakes Region}
\description{Hydrology lines shapefile for Great Lakes Region}
\usage{data(shape_hydrolineClip)}
\keyword{datasets}
\keyword{shapefile}
\keyword{lines}
\keyword{Great Lakes}

