\name{shape_poliboundsClip}
\alias{shape_poliboundsClip}
\docType{data}
\title{Political boundaries polygon shapefile for Great Lakes Region}
\description{Political boundaries polygon shapefile for Great Lakes Region}
\usage{data(shape_poliboundsClip)}
\keyword{datasets}
\keyword{shapefile}
\keyword{polygons}
\keyword{Great Lakes}

