\name{subShape_hydropolyClip}
\alias{subShape_hydropolyClip}
\docType{data}
\title{Hydrology polygon shapefile for Great Lakes Region}
\description{Hydrology polygon shapefile for Great Lakes Region}
\usage{data(subShape_hydropolyClip)}
\keyword{datasets}
\keyword{shapefile}
\keyword{polygons}
\keyword{Great Lakes}

